FIXED Sidebar menu selection
